+++ build-bk/client/assets/PrimaryButton-BpCH6jRa.js
@@ -1,1 +1,0 @@
-import{j as r}from"./chunk-UH6JLGW7-CryfoZLB.js";import{B as i}from"./notifications-BHxKZRbx.js";function s({children:t="Button text",...o}){return r.jsx(i,{className:"!bg-gold w-fit !px-10 !py-1.5 !h-fit !rounded-full transition-all duration-300 active:scale-95 !text-sm hover:!bg-opacity-50 !min-w-[100px]",...o,children:t})}export{s as P};