+++ build-bk/client/assets/useRedirect-Bc6-br2M.js
@@ -1,1 +1,0 @@
-import{u as r,r as s}from"./chunk-UH6JLGW7-CryfoZLB.js";const u=(e,a="/")=>{const t=r();s.useEffect(()=>{e&&t(a,{replace:!0})},[e,t,a])};export{u};