+++ build-bk/client/assets/webRoutes-BaJYf4n_.js
@@ -1,1 +1,0 @@
-const e={bookmarks:"/co/bookmarks",companies:"/companies",feed:"/",market:"/market",messages:"/messages",representatives:"/co/representatives",settings:"/co/settings"};export{e as w};