+++ src/pages/AppLayout.jsx
@@ -1,5 +1,10 @@
 import clsx from "clsx";
-import { Outlet, useLocation, useNavigation } from "react-router";
+import {
+  isRouteErrorResponse,
+  Outlet,
+  useLocation,
+  useNavigation,
+} from "react-router";
 import Sidebar from "../components/admin/markets/sidebar";
 import Navbar from "../components/userProfile/Navbar";
 
@@ -44,3 +49,32 @@ const AppLayout = () => {
 };
 
 export default AppLayout;
+
+export function ErrorBoundary({ error }) {
+  let message = "Oops!";
+  let details = "An unexpected error occurred.";
+  let stack = undefined;
+
+  if (isRouteErrorResponse(error)) {
+    message = error.status === 404 ? "404" : "Error";
+    details =
+      error.status === 404
+        ? "The requested page could not be found."
+        : error.statusText || details;
+  } else if (import.meta.env.DEV && error && error instanceof Error) {
+    details = error.message;
+    stack = error.stack;
+  }
+
+  return (
+    <main className="pt-16 p-4 container mx-auto">
+      <h1>{message}</h1>
+      <p>{details}</p>
+      {stack && (
+        <pre className="w-full p-4 overflow-x-auto">
+          <code>{stack}</code>
+        </pre>
+      )}
+    </main>
+  );
+}