+++ tailwind.config.js
@@ -8,7 +8,7 @@ module.exports = {
   content: [
     "./index.html",
     "./src/**/*.{js,jsx,ts,tsx}",
-    "app/**/*.{ts,tsx}",
+    "app/**/*.{js,jsx,ts,tsx}",
     "components/**/*.{js,jsx,ts,tsx}",
   ],
   theme: {